// src/correios-api/dto/response-price-correios.dto.ts
export interface ResponsePriceCorreiosDTO {
  status: number;
  mensagemPrecoAgencia: string;
  prazo: string;
  url: string;
  mensagemPrecoPPN: string;
  codProdutoAgencia: string;
  precoPPN: string;
  codProdutoPPN: string;
  mensagemPrazo: string;
  msg: string;
  precoAgencia: string;
  urlTitulo: string;
}